requirements.txt
transformers~=4.46.0
easyocr~=1.7.2
numpy~=2.1.2
opencv-python~=4.10.0.84
Flask~=3.0.3
happytransformer~=3.0.0
accelerate~=0.34.2
torch~=2.5.1+cu124
Werkzeug~=3.0.6
python-docx~=1.1.2
pillow~=11.0.0
matplotlib~=3.9.2
pandas~=2.2.3
evaluate~=0.4.3
tqdm~=4.66.6
datasets~=2.21.0
torchmetrics~=1.7.0
torchvision~=0.20.1+cu124